Consolidate fixed-width integer types into a single integer type
The eight named integer types encoded both bit-width and signedness into
the type name, which is ambiguous (storage layout vs. value range vs.
implied signedness), assumes computer-science fluency, and is inconsistent
with every other bounded primitive family in OOML, where the base type
carries separate constraint properties. Range and signedness are now
expressed exclusively through the existing minimum/maximum properties,
which also gives OOML an unbounded integer for the first time.

float32/float64 are untouched: precision affects rounding across the whole
value range, not only at the boundary, so the named-variant pattern earns
its place there in a way it does not for integers.

- OOML.md §6: one integer row replaces the eight, plus a note that the
  bounds describe a portable value range, not a representation
- OOML.md §5.4/§5.5: numeric-bound narrowing is MAJOR and widening MINOR,
  replacing the int32 to int64 type-widening example
- OOML.md §17, README.md: weeklyHours example uses integer with bounds
- schema.json: primitiveTypeName enum
- utils/schema-to-ooml.ts: schema.org Integer maps to integer

Closes #2

+++ utils/schema-to-ooml.ts
@@ -150,7 +150,7 @@ const SCHEMA_DATATYPE_MAP: Record<string, string> = {
     Text: "string",
     Boolean: "boolean",
     Number: "float64",
-    Integer: "int64",
+    Integer: "integer",
     Float: "float64",
     Date: "date",
     Time: "time",
